reverting to the older filter generation file
