+++ src/auth-service/utils/common/generate-filter.util.js
@@ -7,7 +7,6 @@ const isEmpty = require("is-empty");
 const logger = log4js.getLogger(
   `${constants.ENVIRONMENT} -- generate-filter-util`
 );
-
 const {
   logObject,
   logText,
@@ -25,891 +24,1241 @@ const {
   addDays,
 } = require("./date.util");
 
-// ====================== VALIDATION STRATEGIES ======================
-class ValidationStrategy {
-  validate(value, fieldName) {
-    throw new Error("Validation strategy must implement validate method");
-  }
-}
-
-class ObjectIdValidationStrategy extends ValidationStrategy {
-  validate(value, fieldName) {
-    if (!value) return null;
-
-    if (typeof value !== "string" && typeof value !== "object") {
-      logger.warn(
-        `Invalid type for ${fieldName}: expected string or ObjectId, got ${typeof value}`
-      );
-      return null;
-    }
-
-    if (mongoose.Types.ObjectId.isValid(value)) {
-      try {
-        return value instanceof ObjectId ? value : new ObjectId(value);
-      } catch (error) {
-        logger.warn(
-          `Failed to create ObjectId for ${fieldName}: ${value} - ${error.message}`
-        );
-        return null;
-      }
-    }
-
-    logger.warn(`Invalid ObjectId format for ${fieldName}: ${value}`);
-    return null;
-  }
-}
-
-class StringValidationStrategy extends ValidationStrategy {
-  validate(value, fieldName) {
-    if (!value) return null;
-    return typeof value === "string" ? value.trim() : String(value).trim();
-  }
-}
-
-class BooleanValidationStrategy extends ValidationStrategy {
-  validate(value, fieldName) {
-    if (value === undefined || value === null) return null;
-    if (typeof value === "boolean") return value;
-    if (typeof value === "string") {
-      const lower = value.toLowerCase();
-      if (lower === "true" || lower === "yes") return true;
-      if (lower === "false" || lower === "no") return false;
-    }
-    return null;
-  }
-}
-
-class NumberValidationStrategy extends ValidationStrategy {
-  validate(value, fieldName) {
-    if (!value) return null;
-    const num = parseInt(value, 10);
-    return isNaN(num) ? null : num;
-  }
-}
-
-// ====================== VALIDATOR FACTORY ======================
-class ValidatorFactory {
-  static strategies = {
-    objectId: new ObjectIdValidationStrategy(),
-    string: new StringValidationStrategy(),
-    boolean: new BooleanValidationStrategy(),
-    number: new NumberValidationStrategy(),
-  };
-
-  static getValidator(type) {
-    const strategy = this.strategies[type];
-    if (!strategy) {
-      throw new Error(`Unknown validation strategy: ${type}`);
-    }
-    return strategy;
-  }
-}
-
-// ====================== FIELD DEFINITION REGISTRY ======================
-class FieldRegistry {
-  // Common fields that should be ignored in filter processing (pagination, etc.)
-  static ignoredFields = new Set([
-    "limit",
-    "skip",
-    "page",
-    "sort",
-    "order",
-    "fields",
-    "populate",
-  ]);
-
-  static fieldDefinitions = {
-    // Common fields used across multiple entities
-    id: { validator: "objectId", mongoField: "_id" },
-    user_id: { validator: "objectId", mongoField: "user_id" },
-    network_id: { validator: "objectId", mongoField: "network_id" },
-    group_id: { validator: "objectId", mongoField: "group_id" },
-    email: { validator: "string", mongoField: "email" },
-    category: { validator: "string", mongoField: "category" },
-    status: { validator: "string", mongoField: "status" },
-
-    // Entity-specific field mappings
-    users: {
-      privilege: { validator: "string", mongoField: "privilege" },
-      username: { validator: "string", mongoField: "userName" },
-      active: {
-        validator: "string",
-        mongoField: "isActive",
-        transform: (value) =>
-          value === "yes" ? true : value === "no" ? false : null,
-      },
-      email_address: { validator: "string", mongoField: "email" },
-      role_id: { validator: "objectId", mongoField: "role" },
-      resetPasswordToken: {
-        validator: "string",
-        mongoField: "resetPasswordToken",
-      },
-      user: { validator: "objectId", mongoField: "_id" },
-      user_id: { validator: "objectId", mongoField: "_id" },
-      login_count: { validator: "number", mongoField: "login_count" },
-    },
-
-    guest_users: {
-      guest_id: { validator: "string", mongoField: "guest_id" },
-    },
-
-    tenantSettings: {
-      tenant: {
-        validator: "string",
-        mongoField: "tenant",
-        transform: (value) => value.toLowerCase(),
-      },
-    },
-
-    networks: {
-      net_id: { validator: "objectId", mongoField: "_id" },
-      net_email: { validator: "string", mongoField: "net_email" },
-      net_category: { validator: "string", mongoField: "net_category" },
-      category: { validator: "string", mongoField: "category" },
-      net_tenant: { validator: "string", mongoField: "net_tenant" },
-      net_status: { validator: "string", mongoField: "net_status" },
-      net_phoneNumber: { validator: "string", mongoField: "net_phoneNumber" },
-      net_website: { validator: "string", mongoField: "net_website" },
-      net_acronym: { validator: "string", mongoField: "net_acronym" },
-    },
-
-    transactions: {
-      paddle_transaction_id: {
-        validator: "string",
-        mongoField: "paddle_transaction_id",
-      },
-      paddle_event_type: {
-        validator: "string",
-        mongoField: "paddle_event_type",
-      },
-      user_id: { validator: "objectId", mongoField: "user_id" },
-      paddle_customer_id: {
-        validator: "string",
-        mongoField: "paddle_customer_id",
-      },
-      amount: { validator: "string", mongoField: "amount" },
-      currency: { validator: "string", mongoField: "currency" },
-      payment_method: { validator: "string", mongoField: "payment_method" },
-      description: { validator: "string", mongoField: "description" },
-      metadata: { validator: "string", mongoField: "metadata" },
-      donation_campaign_id: {
-        validator: "objectId",
-        mongoField: "donation_campaign_id",
-      },
-    },
-
-    campaigns: {
-      title: {
-        validator: "string",
-        mongoField: "title",
-        transform: (value) => ({ $regex: value, $options: "i" }),
-      },
-      description: {
-        validator: "string",
-        mongoField: "description",
-        transform: (value) => ({ $regex: value, $options: "i" }),
-      },
-      created_by: { validator: "objectId", mongoField: "created_by" },
-      target_amount: { validator: "string", mongoField: "target_amount" },
-      current_amount: { validator: "string", mongoField: "current_amount" },
-      currency: {
-        validator: "string",
-        mongoField: "currency",
-        transform: (value) => value.toUpperCase(),
-      },
-      start_date: {
-        validator: "string",
-        mongoField: "start_date",
-        transform: (value) => ({ $gte: new Date(value) }),
-      },
-      end_date: {
-        validator: "string",
-        mongoField: "end_date",
-        transform: (value) => ({ $lte: new Date(value) }),
-      },
-      is_public: { validator: "boolean", mongoField: "is_public" },
-      tags: {
-        validator: "string",
-        mongoField: "tags",
-        transform: (value) => (Array.isArray(value) ? { $all: value } : value),
-      },
-    },
-
-    candidates: {
-      email_address: { validator: "string", mongoField: "email" },
-      network_id: { validator: "objectId", mongoField: "network_id" },
-    },
-
-    requests: {
-      request_id: { validator: "objectId", mongoField: "_id" },
-      user_id: { validator: "objectId", mongoField: "user_id" },
-      requestType: { validator: "string", mongoField: "requestType" },
-      targetId: { validator: "objectId", mongoField: "targetId" },
-      grp_id: { validator: "objectId", mongoField: "targetId" },
-      net_id: { validator: "objectId", mongoField: "targetId" },
-    },
-
-    defaults: {
-      user: { validator: "objectId", mongoField: "user" },
-      user_id: { validator: "objectId", mongoField: "user" },
-      site: { validator: "objectId", mongoField: "site" },
-      airqloud: { validator: "objectId", mongoField: "airqloud" },
-      grid: { validator: "objectId", mongoField: "grid" },
-      cohort: { validator: "objectId", mongoField: "cohort" },
-      group_id: { validator: "objectId", mongoField: "group_id" },
-      network_id: { validator: "objectId", mongoField: "network_id" },
-    },
-
-    preferences: {
-      user_id: { validator: "objectId", mongoField: "user_id" },
-      group_id: { validator: "objectId", mongoField: "group_id" },
-    },
-
-    maintenances: {
-      product: { validator: "string", mongoField: "product" },
-      tenant: { validator: "string", mongoField: "tenant" },
-    },
-
-    selected_sites: {
-      site_id: { validator: "string", mongoField: "site_id" },
-    },
-
-    checklists: {
-      user_id: { validator: "objectId", mongoField: "user_id" },
-    },
-
-    inquiry: {
-      // Note: In original, 'id' is NOT converted to ObjectId - keeping as string
-      id: { validator: "string", mongoField: "_id" },
-    },
-
-    roles: {
-      role_id: { validator: "objectId", mongoField: "_id" },
-      role_name: { validator: "string", mongoField: "role_name" },
-      role_code: { validator: "string", mongoField: "role_code" },
-      network_id: { validator: "objectId", mongoField: "network_id" },
-      net_id: { validator: "objectId", mongoField: "network_id" },
-      grp_id: { validator: "objectId", mongoField: "group_id" },
-      group_id: { validator: "objectId", mongoField: "group_id" },
-      role_status: { validator: "string", mongoField: "role_status" },
-    },
-
-    permissions: {
-      permission_id: { validator: "string", mongoField: "permission" },
-      network: { validator: "objectId", mongoField: "network_id" },
-      network_id: { validator: "objectId", mongoField: "network_id" },
-      permission: { validator: "string", mongoField: "permission" },
-    },
-
-    tokens: {
-      id: { validator: "objectId", mongoField: "_id" },
-      token: { validator: "string", mongoField: "token" },
-      client_id: { validator: "objectId", mongoField: "client_id" },
-      name: { validator: "string", mongoField: "name" },
-      tenant: { validator: "string", mongoField: "tenant" },
-      emailed: {
-        validator: "string",
-        mongoField: "expiredEmailSent",
-        transform: (value) => value.toLowerCase() === "yes",
-      },
-    },
-
-    activities: {
-      tenant: { validator: "string", mongoField: "tenant" },
-      service: { validator: "string", mongoField: "service" },
-      startTime: { validator: "string", mongoField: "startTime" },
-      endTime: { validator: "string", mongoField: "endTime" },
-      email: { validator: "string", mongoField: "email" },
-    },
-
-    ips: {
-      ip: { validator: "string", mongoField: "ip" },
-      range: { validator: "string", mongoField: "range" },
-      prefix: { validator: "string", mongoField: "prefix" },
-    },
-
-    clients: {
-      client_id: { validator: "objectId", mongoField: "_id" },
-      user_id: { validator: "objectId", mongoField: "user_id" },
-      client_secret: { validator: "string", mongoField: "client_secret" },
-    },
-
-    scopes: {
-      scope_id: { validator: "string", mongoField: "scope" },
-      network_id: { validator: "objectId", mongoField: "network_id" },
-      scope: { validator: "string", mongoField: "scope" },
-    },
-
-    departments: {
-      dep_id: { validator: "objectId", mongoField: "_id" },
-      dep_status: { validator: "string", mongoField: "net_status" },
-      dep_network_id: { validator: "objectId", mongoField: "dep_network_id" },
-      dep_children: { validator: "string", mongoField: "dep_children" },
-      user_id: { validator: "objectId", mongoField: "user_id" },
-      dep_email: { validator: "string", mongoField: "dep_email" },
-      dep_title: { validator: "string", mongoField: "dep_title" },
-      dep_parent: { validator: "objectId", mongoField: "dep_parent" },
-      dep_manager: { validator: "objectId", mongoField: "dep_manager" },
-      has_children: { validator: "boolean", mongoField: "has_children" },
-      dep_acronym: { validator: "string", mongoField: "dep_acronym" },
-    },
-
-    groups: {
-      grp_id: { validator: "objectId", mongoField: "_id" },
-      grp_title: { validator: "string", mongoField: "grp_title" },
-      grp_status: { validator: "string", mongoField: "grp_status" },
-    },
-
-    favorites: {
-      favorite_id: { validator: "objectId", mongoField: "_id" },
-      firebase_user_id: { validator: "string", mongoField: "firebase_user_id" },
-    },
-
-    location_histories: {
-      location_history_id: { validator: "objectId", mongoField: "_id" },
-      firebase_user_id: { validator: "string", mongoField: "firebase_user_id" },
-    },
-
-    search_histories: {
-      search_history_id: { validator: "objectId", mongoField: "_id" },
-      firebase_user_id: { validator: "string", mongoField: "firebase_user_id" },
-    },
-  };
-
-  static isIgnoredField(fieldName) {
-    return this.ignoredFields.has(fieldName);
-  }
-
-  static getFieldDefinition(entity, field) {
-    // Check entity-specific definitions first
-    if (this.fieldDefinitions[entity] && this.fieldDefinitions[entity][field]) {
-      return this.fieldDefinitions[entity][field];
-    }
-    // Fall back to common field definitions
-    return this.fieldDefinitions[field] || null;
-  }
-}
-
-// ====================== BASE FILTER BUILDER ======================
-class BaseFilterBuilder {
-  constructor(entityType) {
-    this.entityType = entityType;
-    this.filter = {};
-    this.errors = [];
-  }
-
-  addField(fieldName, value) {
-    if (value === undefined || value === null || value === "") {
-      return this;
-    }
-
-    // Skip ignored fields (pagination parameters) early
-    if (FieldRegistry.isIgnoredField(fieldName)) {
-      return this;
-    }
-
-    const fieldDef = FieldRegistry.getFieldDefinition(
-      this.entityType,
-      fieldName
-    );
-    if (!fieldDef) {
-      // FIXED: Simplified warning logic - removed redundant check
-      logger.warn(
-        `No field definition found for ${this.entityType}.${fieldName}`
-      );
-      return this;
-    }
-
+const filter = {
+  users: (req, next) => {
     try {
-      const validator = ValidatorFactory.getValidator(fieldDef.validator);
-      const validatedValue = validator.validate(value, fieldName);
+      let {
+        privilege,
+        id,
+        username,
+        active,
+        email_address,
+        role_id,
+        email,
+        resetPasswordToken,
+        user,
+        user_id,
+        login_count,
+      } = { ...req.body, ...req.query, ...req.params };
 
-      if (validatedValue !== null) {
-        const mongoField = fieldDef.mongoField || fieldName;
-        const finalValue = fieldDef.transform
-          ? fieldDef.transform(validatedValue)
-          : validatedValue;
-        this.filter[mongoField] = finalValue;
+      let filter = {};
+
+      if (email) {
+        filter["email"] = email;
       }
+
+      if (role_id) {
+        filter["role"] = ObjectId(role_id);
+      }
+
+      if (email_address) {
+        filter["email"] = email_address;
+      }
+      if (!isEmpty(resetPasswordToken)) {
+        filter["resetPasswordToken"] = resetPasswordToken;
+      }
+      if (privilege) {
+        filter["privilege"] = privilege;
+      }
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+      if (user_id) {
+        filter["_id"] = ObjectId(user_id);
+      }
+      if (user) {
+        filter["_id"] = ObjectId(user);
+      }
+      if (active) {
+        if (active === "yes") {
+          filter["isActive"] = true;
+        } else if (active === "no") {
+          filter["isActive"] = false;
+        }
+      }
+      if (username) {
+        filter["userName"] = username;
+      }
+
+      if (login_count) {
+        filter["login_count"] = parseInt(login_count);
+      }
+
+      return filter;
     } catch (error) {
-      this.errors.push(`Error processing field ${fieldName}: ${error.message}`);
-      logger.error(`Error processing field ${fieldName}: ${error.message}`);
-    }
-
-    return this;
-  }
-
-  addMultipleFields(data) {
-    Object.entries(data).forEach(([key, value]) => {
-      this.addField(key, value);
-    });
-    return this;
-  }
-
-  build() {
-    if (this.errors.length > 0) {
-      logger.warn(`Filter built with errors: ${this.errors.join(", ")}`);
-    }
-    return this.filter;
-  }
-
-  hasErrors() {
-    return this.errors.length > 0;
-  }
-
-  getErrors() {
-    return this.errors;
-  }
-}
-
-// ====================== SPECIALIZED FILTER BUILDERS ======================
-class DateRangeFilterBuilder extends BaseFilterBuilder {
-  addDateRange(startField, endField, startValue, endValue, next) {
-    const today = monthsInfront(0, next);
-    const oneWeekBack = addDays(-7, next);
-
-    let dateFilter = {
-      $gte: oneWeekBack,
-      $lte: today,
-    };
-
-    if (startValue && !endValue) {
-      if (!isTimeEmpty(startValue)) {
-        dateFilter.$gte = addMonthsToProvideDateTime(startValue, 1, next);
-      } else {
-        dateFilter.$gte = addMonthsToProvidedDate(startValue, 1, next);
-      }
-    } else if (endValue && !startValue) {
-      if (!isTimeEmpty(endValue)) {
-        dateFilter.$lte = addMonthsToProvideDateTime(endValue, -1, next);
-      } else {
-        dateFilter.$lte = addMonthsToProvidedDate(endValue, -1, next);
-      }
-    } else if (endValue && startValue) {
-      dateFilter.$lte = new Date(endValue);
-      dateFilter.$gte = new Date(startValue);
-    }
-
-    this.filter[startField || "timestamp"] = dateFilter;
-    return this;
-  }
-}
-
-class ActivityFilterBuilder extends DateRangeFilterBuilder {
-  addTenantValidation(tenant) {
-    if (!tenant) {
-      this.errors.push("Tenant is required for activity filters");
-      return this;
-    }
-    this.filter.tenant = tenant;
-    return this;
-  }
-
-  addServiceFilter(service, dateFilter) {
-    if (!service) return this;
-
-    const serviceMatch = { $elemMatch: { name: service } };
-
-    if (Object.keys(dateFilter || {}).length > 0) {
-      this.filter.dailyStats = {
-        $elemMatch: {
-          date: dateFilter,
-          services: serviceMatch,
-        },
-      };
-    } else {
-      this.filter.dailyStats = {
-        $elemMatch: {
-          services: serviceMatch,
-        },
-      };
-    }
-
-    this.filter.monthlyStats = {
-      $elemMatch: {
-        uniqueServices: service,
-      },
-    };
-
-    return this;
-  }
-}
-
-class SearchFilterBuilder extends BaseFilterBuilder {
-  addTextSearch(field, value, options = { caseInsensitive: true }) {
-    if (!value) return this;
-
-    if (options.caseInsensitive) {
-      this.filter[field] = { $regex: value, $options: "i" };
-    } else {
-      this.filter[field] = { $regex: value };
-    }
-    return this;
-  }
-
-  addArrayFilter(field, value, matchAll = false) {
-    if (!value) return this;
-
-    if (Array.isArray(value)) {
-      this.filter[field] = matchAll ? { $all: value } : { $in: value };
-    } else {
-      this.filter[field] = value;
-    }
-    return this;
-  }
-}
-
-// ====================== FILTER FACTORY ======================
-class FilterFactory {
-  static builderMap = {
-    users: BaseFilterBuilder,
-    guest_users: BaseFilterBuilder,
-    tenantSettings: BaseFilterBuilder,
-    networks: BaseFilterBuilder,
-    transactions: BaseFilterBuilder,
-    campaigns: SearchFilterBuilder,
-    candidates: BaseFilterBuilder,
-    requests: BaseFilterBuilder,
-    defaults: BaseFilterBuilder,
-    preferences: BaseFilterBuilder,
-    maintenances: BaseFilterBuilder,
-    selected_sites: BaseFilterBuilder,
-    checklists: BaseFilterBuilder,
-    inquiry: BaseFilterBuilder,
-    roles: BaseFilterBuilder,
-    permissions: BaseFilterBuilder,
-    tokens: BaseFilterBuilder,
-    ips: BaseFilterBuilder,
-    clients: BaseFilterBuilder,
-    scopes: BaseFilterBuilder,
-    departments: BaseFilterBuilder,
-    groups: BaseFilterBuilder,
-    logs: DateRangeFilterBuilder,
-    activities: ActivityFilterBuilder,
-    favorites: BaseFilterBuilder,
-    location_histories: BaseFilterBuilder,
-    search_histories: BaseFilterBuilder,
-  };
-
-  static createBuilder(entityType) {
-    const BuilderClass = this.builderMap[entityType] || BaseFilterBuilder;
-    return new BuilderClass(entityType);
-  }
-}
-
-// ====================== MAIN FILTER PROCESSOR ======================
-class FilterProcessor {
-  static extractRequestData(req) {
-    return { ...req.body, ...req.query, ...req.params };
-  }
-
-  static createErrorHandler(next) {
-    return (error) => {
       logger.error(`🐛🐛 Internal Server Error ${error.message}`);
-      if (next) {
-        next(
-          new HttpError(
-            "Internal Server Error",
-            httpStatus.INTERNAL_SERVER_ERROR,
-            { message: error.message }
-          )
-        );
-      }
-    };
-  }
-
-  static processFilter(entityType, req, next, customProcessor = null) {
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  guest_users: (req, next) => {
     try {
-      const data = this.extractRequestData(req);
-      const builder = FilterFactory.createBuilder(entityType);
-
-      if (customProcessor) {
-        return customProcessor(builder, data, req, next);
+      let { id, guest_id } = {
+        ...req.body,
+        ...req.query,
+        ...req.params,
+      };
+      let filter = {};
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+      if (guest_id) {
+        filter["guest_id"] = guest_id;
       }
 
-      const filter = builder.addMultipleFields(data).build();
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  tenantSettings: (req, next) => {
+    try {
+      const { query, params } = req;
+      const { tenant, id } = { ...query, ...params };
 
-      if (builder.hasErrors()) {
-        logger.warn(
-          `Filter errors for ${entityType}: ${builder.getErrors().join(", ")}`
-        );
-        // For critical entities like tokens, log more details
-        if (entityType === "tokens") {
-          logger.error(
-            `🐛🐛 Critical filter errors for ${entityType}: ${builder
-              .getErrors()
-              .join(", ")}`
-          );
+      let filter = {};
+      if (tenant) {
+        filter.tenant = tenant.toLowerCase();
+      }
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  networks: (req, next) => {
+    try {
+      const {
+        net_id,
+        net_email,
+        net_category,
+        net_tenant,
+        net_status,
+        net_phoneNumber,
+        net_website,
+        net_acronym,
+        category,
+      } = { ...req.query, ...req.params };
+
+      let filter = {};
+      if (net_email) {
+        filter["net_email"] = net_email;
+      }
+      if (category) {
+        filter["category"] = category;
+      }
+
+      if (net_category) {
+        filter["net_category"] = net_category;
+      }
+
+      if (net_id) {
+        filter["_id"] = ObjectId(net_id);
+      }
+
+      if (net_tenant) {
+        filter["net_tenant"] = net_tenant;
+      }
+      if (net_acronym) {
+        filter["net_acronym"] = net_acronym;
+      }
+
+      if (net_phoneNumber) {
+        filter["net_phoneNumber"] = net_phoneNumber;
+      }
+      if (net_website) {
+        filter["net_website"] = net_website;
+      }
+      if (net_status) {
+        filter["net_status"] = net_status;
+      }
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  transactions: (req, next) => {
+    try {
+      const {
+        paddle_transaction_id,
+        paddle_event_type,
+        user_id,
+        paddle_customer_id,
+        amount,
+        currency,
+        payment_method,
+        status,
+        description,
+        metadata,
+        donation_campaign_id,
+      } = { ...req.query, ...req.params };
+
+      let filter = {};
+
+      if (paddle_transaction_id) {
+        filter["paddle_transaction_id"] = paddle_transaction_id;
+      }
+
+      if (paddle_event_type) {
+        filter["paddle_event_type"] = paddle_event_type;
+      }
+
+      if (user_id) {
+        filter["user_id"] = ObjectId(user_id);
+      }
+
+      if (paddle_customer_id) {
+        filter["paddle_customer_id"] = paddle_customer_id;
+      }
+
+      if (amount) {
+        filter["amount"] = amount;
+      }
+
+      if (currency) {
+        filter["currency"] = currency;
+      }
+
+      if (payment_method) {
+        filter["payment_method"] = payment_method;
+      }
+
+      if (status) {
+        filter["status"] = status;
+      }
+
+      if (description) {
+        filter["description"] = description;
+      }
+
+      if (metadata) {
+        filter["metadata"] = metadata;
+      }
+
+      if (donation_campaign_id) {
+        filter["donation_campaign_id"] = ObjectId(donation_campaign_id);
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  campaigns: (req, next) => {
+    try {
+      const {
+        title,
+        description,
+        created_by,
+        target_amount,
+        current_amount,
+        currency,
+        start_date,
+        end_date,
+        status,
+        is_public,
+        category,
+        tags,
+      } = { ...req.query, ...req.params, ...req.body }; // Include req.body
+
+      let filter = {};
+
+      if (title) {
+        filter.title = { $regex: title, $options: "i" }; // Case-insensitive search
+      }
+
+      if (description) {
+        filter.description = { $regex: description, $options: "i" };
+      }
+
+      if (created_by) {
+        filter.created_by = ObjectId(created_by);
+      }
+
+      if (target_amount) {
+        filter.target_amount = target_amount;
+      }
+
+      if (current_amount) {
+        filter.current_amount = current_amount;
+      }
+
+      if (currency) {
+        filter.currency = currency.toUpperCase(); // Ensure consistent case
+      }
+
+      if (start_date) {
+        filter.start_date = { $gte: new Date(start_date) }; // Greater than or equal
+      }
+
+      if (end_date) {
+        filter.end_date = { $lte: new Date(end_date) }; // Less than or equal
+      }
+
+      if (status) {
+        filter.status = status;
+      }
+
+      if (is_public !== undefined) {
+        // Handle boolean values correctly
+        filter.is_public = is_public;
+      }
+
+      if (category) {
+        filter.category = category;
+      }
+
+      if (tags) {
+        // Check if tags are provided
+        if (Array.isArray(tags)) {
+          //If an array of tags is provided
+          filter.tags = { $all: tags }; // Match all tags in the array
+        } else if (typeof tags === "string") {
+          //If a single tag string is provided
+          filter.tags = tags; // Match the single tag
         }
       }
 
       return filter;
     } catch (error) {
-      logger.error(
-        `🐛🐛 Filter processing error for ${entityType}: ${error.message}`
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
       );
-      this.createErrorHandler(next)(error);
-      return {};
     }
-  }
-}
-
-// ====================== CUSTOM PROCESSORS ======================
-class CustomProcessors {
-  static logs(builder, data, req, next) {
-    const { service, startTime, endTime, email } = data;
-    const today = monthsInfront(0, next);
-    const oneWeekBack = addDays(-7, next);
-
-    let filter = {
-      timestamp: {
-        $gte: oneWeekBack,
-        $lte: today,
-      },
-    };
-
-    if (service) {
-      logText("service present ");
-      filter["meta.service"] = service;
-    }
-
-    if (startTime && isEmpty(endTime)) {
-      logText("startTime present and endTime is missing");
-      if (isTimeEmpty(startTime) === false) {
-        filter["timestamp"]["$gte"] = addMonthsToProvideDateTime(
-          startTime,
-          1,
-          next
-        );
-      } else {
-        filter["timestamp"]["$gte"] = addMonthsToProvidedDate(
-          startTime,
-          1,
-          next
-        );
-      }
-    } else if (endTime && isEmpty(startTime)) {
-      logText("startTime absent and endTime is present");
-      if (isTimeEmpty(endTime) === false) {
-        filter["timestamp"]["$lte"] = addMonthsToProvideDateTime(
-          endTime,
-          -1,
-          next
-        );
-      } else {
-        filter["timestamp"]["$lte"] = addMonthsToProvidedDate(
-          endTime,
-          -1,
-          next
-        );
-      }
-    } else if (endTime && startTime) {
-      logText("startTime present and endTime is also present");
-      filter["timestamp"]["$lte"] = new Date(endTime);
-      filter["timestamp"]["$gte"] = new Date(startTime);
-    }
-
-    if (email) {
-      logText("email present ");
-      filter["meta.email"] = email;
-    }
-
-    return filter;
-  }
-
-  static activities(builder, data, req, next) {
-    const { service, startTime, endTime, email, tenant } = data;
-    const today = monthsInfront(0, next);
-    const oneWeekBack = addDays(-7, next);
-
-    logObject("the req.query in the activity filter", req.query);
-
-    // Base filter
-    let filter = {};
-
-    // Handle required tenant field
-    if (!tenant) {
-      throw new Error("Tenant is required");
-    }
-    filter.tenant = tenant;
-
-    // Handle email filtering
-    if (email) {
-      logText("email present");
-      filter.email = email;
-    }
-
-    // Date range handling for dailyStats
-    let dateFilter = {};
-    if (startTime && !endTime) {
-      logText("startTime present and endTime is missing");
-      if (!isTimeEmpty(startTime)) {
-        dateFilter.$gte = addMonthsToProvideDateTime(startTime, 1, next);
-      } else {
-        dateFilter.$gte = addMonthsToProvidedDate(startTime, 1, next);
-      }
-    } else if (endTime && !startTime) {
-      logText("startTime absent and endTime is present");
-      if (!isTimeEmpty(endTime)) {
-        dateFilter.$lte = addMonthsToProvideDateTime(endTime, -1, next);
-      } else {
-        dateFilter.$lte = addMonthsToProvidedDate(endTime, -1, next);
-      }
-    } else if (endTime && startTime) {
-      logText("startTime present and endTime is also present");
-      dateFilter.$lte = new Date(endTime);
-      dateFilter.$gte = new Date(startTime);
-    } else {
-      // Default to one week range
-      dateFilter.$gte = oneWeekBack;
-      dateFilter.$lte = today;
-    }
-
-    // Apply date filter to dailyStats
-    if (Object.keys(dateFilter).length > 0) {
-      filter["dailyStats"] = {
-        $elemMatch: {
-          date: dateFilter,
-        },
+  },
+  candidates: (req, next) => {
+    try {
+      let { category, id, email_address, email, network_id } = {
+        ...req.body,
+        ...req.query,
+        ...req.params,
       };
+      let filter = {};
+      if (email) {
+        filter["email"] = email;
+      }
+      if (network_id) {
+        filter["network_id"] = ObjectId(network_id);
+      }
+      if (email_address) {
+        filter["email"] = email_address;
+      }
+      if (category) {
+        filter["category"] = category;
+      }
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
     }
+  },
+  requests: (req, next) => {
+    try {
+      const {
+        id,
+        user_id,
+        requestType,
+        targetId,
+        status,
+        category,
+        request_id,
+        grp_id,
+        net_id,
+      } = {
+        ...req.body,
+        ...req.query,
+        ...req.params,
+      };
 
-    // Service filtering
-    if (service) {
-      logText("service present");
-      // Add service filter to the dailyStats elemMatch
-      if (filter.dailyStats && filter.dailyStats.$elemMatch) {
-        filter.dailyStats.$elemMatch["services"] = {
-          $elemMatch: { name: service },
-        };
-      } else {
-        filter.dailyStats = {
+      let filter = {};
+      if (user_id) {
+        filter["user_id"] = ObjectId(user_id);
+      }
+      if (requestType) {
+        filter["requestType"] = requestType;
+      }
+      if (targetId) {
+        filter["targetId"] = ObjectId(targetId);
+      }
+
+      if (grp_id) {
+        filter["targetId"] = ObjectId(grp_id);
+      }
+
+      if (net_id) {
+        filter["targetId"] = ObjectId(net_id);
+      }
+
+      if (status) {
+        filter["status"] = status;
+      }
+      if (category) {
+        filter["category"] = category;
+      }
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      if (request_id) {
+        filter["_id"] = ObjectId(request_id);
+      }
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  defaults: (req, next) => {
+    try {
+      let {
+        id,
+        user,
+        user_id,
+        site,
+        airqloud,
+        grid,
+        cohort,
+        group_id,
+        network_id,
+      } = {
+        ...req.query,
+        ...req.params,
+      };
+      let filter = {};
+      if (user) {
+        filter["user"] = ObjectId(user);
+      }
+
+      if (user_id) {
+        filter["user"] = ObjectId(user_id);
+      }
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      if (grid) {
+        filter["grid"] = ObjectId(grid);
+      }
+
+      if (cohort) {
+        filter["cohort"] = ObjectId(cohort);
+      }
+
+      if (group_id) {
+        filter["group_id"] = ObjectId(group_id);
+      }
+
+      if (network_id) {
+        filter["network_id"] = ObjectId(network_id);
+      }
+
+      if (site) {
+        filter["site"] = ObjectId(site);
+      }
+
+      if (airqloud) {
+        filter["airqloud"] = ObjectId(airqloud);
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  preferences: (req, next) => {
+    try {
+      let { user_id, group_id } = {
+        ...req.body,
+        ...req.query,
+        ...req.params,
+      };
+
+      let filter = {};
+
+      if (user_id) {
+        filter["user_id"] = ObjectId(user_id);
+      }
+
+      if (group_id) {
+        filter["group_id"] = ObjectId(group_id);
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  maintenances: (req, next) => {
+    try {
+      let { id, product } = {
+        ...req.body,
+        ...req.query,
+        ...req.params,
+      };
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      if (product) {
+        filter["product"] = product;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  selected_sites: (req, next) => {
+    try {
+      let { site_id } = {
+        ...req.body,
+        ...req.query,
+        ...req.params,
+      };
+      let filter = {};
+      if (site_id) {
+        filter["site_id"] = site_id;
+      }
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  checklists: (req, next) => {
+    try {
+      let { id, user_id } = {
+        ...req.body,
+        ...req.query,
+        ...req.params,
+      };
+      let filter = {};
+
+      if (user_id) {
+        filter["user_id"] = ObjectId(user_id);
+      }
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  inquiry: (req, next) => {
+    try {
+      let { category, id, email } = {
+        ...req.body,
+        ...req.query,
+        ...req.params,
+      };
+
+      let filter = {};
+      if (email) {
+        filter["email"] = email;
+      }
+      if (category) {
+        filter["category"] = category;
+      }
+      if (id) {
+        filter["_id"] = id;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  roles: (req, next) => {
+    try {
+      const { query, params } = req;
+      const {
+        id,
+        role_id,
+        role_name,
+        role_code,
+        network_id,
+        net_id,
+        grp_id,
+        group_id,
+        role_status,
+        category,
+      } = { ...params, ...query };
+
+      const filter = {};
+
+      if (id || role_id) {
+        filter["_id"] = ObjectId(id || role_id);
+      }
+      if (network_id || net_id) {
+        filter["network_id"] = ObjectId(network_id || net_id);
+      }
+      if (grp_id || group_id) {
+        filter["group_id"] = ObjectId(grp_id || group_id);
+      }
+      if (category) {
+        filter["category"] = category;
+      }
+      if (role_name) {
+        filter["role_name"] = role_name;
+      }
+      if (role_code) {
+        filter["role_code"] = role_code;
+      }
+      if (role_status) {
+        filter["role_status"] = role_status;
+      }
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  permissions: (req, next) => {
+    try {
+      const { query, params } = req;
+      const { id, network, permission } = query;
+      const { permission_id, network_id } = params;
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      if (permission_id) {
+        filter["permission"] = permission_id;
+      }
+
+      if (network) {
+        filter["network_id"] = ObjectId(network);
+      }
+
+      if (network_id) {
+        filter["network_id"] = ObjectId(network_id);
+      }
+
+      if (permission) {
+        filter["permission"] = permission;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  tokens: (req, next) => {
+    try {
+      const { query, params } = req;
+      const { token, client_id, name, id, emailed } = { ...query, ...params };
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      if (token) {
+        filter["token"] = token;
+      }
+
+      if (emailed) {
+        filter.expiredEmailSent = emailed.toLowerCase() === "yes";
+      }
+
+      if (client_id) {
+        filter["client_id"] = ObjectId(client_id);
+      }
+
+      if (name) {
+        filter["name"] = name;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  ips: (req, next) => {
+    try {
+      const { id, ip, range, ip_counts, prefix, prefix_counts } = {
+        ...req.query,
+        ...req.params,
+      };
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      if (ip) {
+        filter["ip"] = ip;
+      }
+
+      if (range) {
+        filter["range"] = range;
+      }
+
+      if (prefix) {
+        filter["prefix"] = prefix;
+      }
+
+      if (ip_counts) {
+        filter["ipCounts"] = {
           $elemMatch: {
-            services: {
-              $elemMatch: { name: service },
+            count: {
+              $gte: parseInt(ip_counts),
             },
           },
         };
       }
 
-      // Also check monthly stats for the service
-      filter["monthlyStats"] = {
-        $elemMatch: {
-          uniqueServices: service,
-        },
-      };
-    }
-
-    return filter;
-  }
-
-  static campaigns(builder, data, req, next) {
-    // The campaigns entity includes body data as well
-    const fullData = { ...req.query, ...req.params, ...req.body };
-
-    // Handle all regular fields
-    builder.addMultipleFields(fullData);
-
-    // Special handling for tags (already defined in field definition, but handling here for safety)
-    const { tags } = fullData;
-    if (tags) {
-      if (Array.isArray(tags)) {
-        builder.filter.tags = { $all: tags };
-      } else if (typeof tags === "string") {
-        builder.filter.tags = tags;
+      if (prefix_counts) {
+        filter["prefixCounts"] = {
+          $elemMatch: {
+            count: {
+              $gte: parseInt(prefix_counts),
+            },
+          },
+        };
       }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
     }
-
-    return builder.build();
-  }
-
-  static ips(builder, data, req, next) {
-    const { ip_counts, prefix_counts } = data;
-
-    // Handle regular fields first
-    builder.addMultipleFields(data);
-
-    // Special handling for count-based queries
-    if (ip_counts) {
-      builder.filter.ipCounts = {
-        $elemMatch: {
-          count: { $gte: parseInt(ip_counts, 10) },
-        },
-      };
-    }
-
-    if (prefix_counts) {
-      builder.filter.prefixCounts = {
-        $elemMatch: {
-          count: { $gte: parseInt(prefix_counts, 10) },
-        },
-      };
-    }
-
-    return builder.build();
-  }
-}
-
-// ====================== PUBLIC API ======================
-const filter = {
-  users: (req, next) => FilterProcessor.processFilter("users", req, next),
-  guest_users: (req, next) =>
-    FilterProcessor.processFilter("guest_users", req, next),
-  tenantSettings: (req, next) =>
-    FilterProcessor.processFilter("tenantSettings", req, next),
-  networks: (req, next) => FilterProcessor.processFilter("networks", req, next),
-  transactions: (req, next) =>
-    FilterProcessor.processFilter("transactions", req, next),
-  campaigns: (req, next) =>
-    FilterProcessor.processFilter(
-      "campaigns",
-      req,
-      next,
-      CustomProcessors.campaigns
-    ),
-  candidates: (req, next) =>
-    FilterProcessor.processFilter("candidates", req, next),
-  requests: (req, next) => FilterProcessor.processFilter("requests", req, next),
-  defaults: (req, next) => FilterProcessor.processFilter("defaults", req, next),
-  preferences: (req, next) =>
-    FilterProcessor.processFilter("preferences", req, next),
-  maintenances: (req, next) =>
-    FilterProcessor.processFilter("maintenances", req, next),
-  selected_sites: (req, next) =>
-    FilterProcessor.processFilter("selected_sites", req, next),
-  checklists: (req, next) =>
-    FilterProcessor.processFilter("checklists", req, next),
-  inquiry: (req, next) => FilterProcessor.processFilter("inquiry", req, next),
-  roles: (req, next) => FilterProcessor.processFilter("roles", req, next),
-  permissions: (req, next) =>
-    FilterProcessor.processFilter("permissions", req, next),
-  tokens: (req, next) => FilterProcessor.processFilter("tokens", req, next),
-  ips: (req, next) =>
-    FilterProcessor.processFilter("ips", req, next, CustomProcessors.ips),
-  clients: (req, next) => FilterProcessor.processFilter("clients", req, next),
-  scopes: (req, next) => FilterProcessor.processFilter("scopes", req, next),
-  departments: (req, next) =>
-    FilterProcessor.processFilter("departments", req, next),
-  groups: (req, next) => {
-    const result = FilterProcessor.processFilter("groups", req, next);
-    logObject("the filter we are sending", result);
-    return result;
   },
-  logs: (req, next) =>
-    FilterProcessor.processFilter("logs", req, next, CustomProcessors.logs),
-  activities: (req, next) =>
-    FilterProcessor.processFilter(
-      "activities",
-      req,
-      next,
-      CustomProcessors.activities
-    ),
-  favorites: (req, next) =>
-    FilterProcessor.processFilter("favorites", req, next),
-  location_histories: (req, next) =>
-    FilterProcessor.processFilter("location_histories", req, next),
-  search_histories: (req, next) =>
-    FilterProcessor.processFilter("search_histories", req, next),
+  clients: (req, next) => {
+    try {
+      const { query, params } = req;
+      const { id, user_id } = query;
+      const { client_id, client_name, network_id, client_secret } = params;
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      if (user_id) {
+        filter["user_id"] = ObjectId(user_id);
+      }
+
+      if (client_id) {
+        filter["_id"] = ObjectId(client_id);
+      }
+
+      if (client_secret) {
+        filter["client_secret"] = client_secret;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  scopes: (req, next) => {
+    try {
+      const { query, params } = req;
+      const { id, scope } = query;
+      const { scope_id, network_id } = params;
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+
+      if (scope_id) {
+        filter["scope"] = scope_id;
+      }
+
+      if (scope) {
+        filter["scope"] = scope;
+      }
+
+      if (network_id) {
+        filter["network_id"] = ObjectId(network_id);
+      }
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  departments: (req, next) => {
+    try {
+      const {
+        dep_id,
+        user_id,
+        dep_email,
+        dep_title,
+        dep_network_id,
+        dep_parent,
+        dep_manager,
+        has_children,
+        dep_acronym,
+        dep_children,
+        dep_status,
+      } = { ...req.query, ...req.params };
+
+      let filter = {};
+
+      if (dep_id) {
+        filter["_id"] = ObjectId(dep_id);
+      }
+
+      if (dep_status) {
+        filter["net_status"] = dep_status;
+      }
+
+      if (dep_network_id) {
+        filter["dep_network_id"] = ObjectId(dep_network_id);
+      }
+
+      if (dep_children) {
+        filter["dep_children"] = dep_children;
+      }
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  groups: (req, next) => {
+    try {
+      const { grp_title, grp_status, category, grp_id } = {
+        ...req.query,
+        ...req.params,
+      };
+      let filter = {};
+
+      if (grp_id) {
+        filter["_id"] = ObjectId(grp_id);
+      }
+      if (grp_status) {
+        filter["grp_status"] = grp_status;
+      }
+      if (grp_title) {
+        filter["grp_title"] = grp_title;
+      }
+
+      if (category) {
+        filter["category"] = category;
+      }
+      logObject("the filter we are sending", filter);
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  logs: (req, next) => {
+    try {
+      const { service, startTime, endTime, email } = req.query;
+      const today = monthsInfront(0, next);
+      const oneWeekBack = addDays(-7, next);
+      logObject("the req.query in the logs filter", req.query);
+
+      let filter = {
+        timestamp: {
+          $gte: oneWeekBack,
+          $lte: today,
+        },
+      };
+
+      if (service) {
+        logText("service present ");
+        filter["meta.service"] = service;
+      }
+
+      if (startTime && isEmpty(endTime)) {
+        logText("startTime present and endTime is missing");
+        if (isTimeEmpty(startTime) === false) {
+          filter["timestamp"]["$gte"] = addMonthsToProvideDateTime(
+            startTime,
+            1,
+            next
+          );
+        } else {
+          filter["timestamp"]["$lte"] = addMonthsToProvidedDate(
+            startTime,
+            1,
+            next
+          );
+        }
+      } else if (endTime && isEmpty(startTime)) {
+        logText("startTime absent and endTime is present");
+        if (isTimeEmpty(endTime) === false) {
+          filter["timestamp"]["$lte"] = addMonthsToProvideDateTime(
+            endTime,
+            -1,
+            next
+          );
+        } else {
+          filter["timestamp"]["$lte"] = addMonthsToProvidedDate(
+            endTime,
+            -1,
+            next
+          );
+        }
+      } else if (endTime && startTime) {
+        logText("startTime present and endTime is also present");
+        filter["timestamp"]["$lte"] = new Date(endTime);
+        filter["timestamp"]["$gte"] = new Date(startTime);
+      }
+
+      if (email) {
+        logText("email present ");
+        filter["meta.email"] = email;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  activities: (req, next) => {
+    try {
+      const { service, startTime, endTime, email, tenant } = req.query;
+      const today = monthsInfront(0, next);
+      const oneWeekBack = addDays(-7, next);
+
+      logObject("the req.query in the activity filter", req.query);
+
+      // Base filter
+      let filter = {};
+
+      // Handle required tenant field
+      if (!tenant) {
+        throw new Error("Tenant is required");
+      }
+      filter.tenant = tenant;
+
+      // Handle email filtering
+      if (email) {
+        logText("email present");
+        filter.email = email;
+      }
+
+      // Date range handling for dailyStats
+      let dateFilter = {};
+      if (startTime && !endTime) {
+        logText("startTime present and endTime is missing");
+        if (!isTimeEmpty(startTime)) {
+          dateFilter.$gte = addMonthsToProvideDateTime(startTime, 1, next);
+        } else {
+          dateFilter.$gte = addMonthsToProvidedDate(startTime, 1, next);
+        }
+      } else if (endTime && !startTime) {
+        logText("startTime absent and endTime is present");
+        if (!isTimeEmpty(endTime)) {
+          dateFilter.$lte = addMonthsToProvideDateTime(endTime, -1, next);
+        } else {
+          dateFilter.$lte = addMonthsToProvidedDate(endTime, -1, next);
+        }
+      } else if (endTime && startTime) {
+        logText("startTime present and endTime is also present");
+        dateFilter.$lte = new Date(endTime);
+        dateFilter.$gte = new Date(startTime);
+      } else {
+        // Default to one week range
+        dateFilter.$gte = oneWeekBack;
+        dateFilter.$lte = today;
+      }
+
+      // Apply date filter to dailyStats
+      if (Object.keys(dateFilter).length > 0) {
+        filter["dailyStats"] = {
+          $elemMatch: {
+            date: dateFilter,
+          },
+        };
+      }
+
+      // Service filtering
+      if (service) {
+        logText("service present");
+        // Add service filter to the dailyStats elemMatch
+        if (filter.dailyStats && filter.dailyStats.$elemMatch) {
+          filter.dailyStats.$elemMatch["services"] = {
+            $elemMatch: { name: service },
+          };
+        } else {
+          filter.dailyStats = {
+            $elemMatch: {
+              services: {
+                $elemMatch: { name: service },
+              },
+            },
+          };
+        }
+
+        // Also check monthly stats for the service
+        filter["monthlyStats"] = {
+          $elemMatch: {
+            uniqueServices: service,
+          },
+        };
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  favorites: (req, next) => {
+    try {
+      const { query, params } = req;
+      const { firebase_user_id, favorite_id, id } = { ...query, ...params };
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+      if (favorite_id) {
+        filter["_id"] = ObjectId(favorite_id);
+      }
+      if (firebase_user_id) {
+        filter["firebase_user_id"] = firebase_user_id;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  location_histories: (req, next) => {
+    try {
+      const { query, params } = req;
+      const { firebase_user_id, location_history_id, id } = {
+        ...query,
+        ...params,
+      };
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+      if (location_history_id) {
+        filter["_id"] = ObjectId(location_history_id);
+      }
+      if (firebase_user_id) {
+        filter["firebase_user_id"] = firebase_user_id;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
+  search_histories: (req, next) => {
+    try {
+      const { query, params } = req;
+      const { firebase_user_id, search_history_id, id } = {
+        ...query,
+        ...params,
+      };
+      let filter = {};
+
+      if (id) {
+        filter["_id"] = ObjectId(id);
+      }
+      if (search_history_id) {
+        filter["_id"] = ObjectId(search_history_id);
+      }
+      if (firebase_user_id) {
+        filter["firebase_user_id"] = firebase_user_id;
+      }
+
+      return filter;
+    } catch (error) {
+      logger.error(`🐛🐛 Internal Server Error ${error.message}`);
+      next(
+        new HttpError(
+          "Internal Server Error",
+          httpStatus.INTERNAL_SERVER_ERROR,
+          { message: error.message }
+        )
+      );
+    }
+  },
 };
 
 module.exports = filter;
